Migrate to Mixed for property access (#2110)
* lay some groundwork for migration to mixed

* Get collections working

* Use primitivevalue for string/byte[]

* After rebase fixes

* Use RealmValue/PrimitiveValue for all types except object

* Revert some accidental changes

* Use PrimitiveValue for Realm.Find

* fix weaver tests

* Add [Preserve] to RealmValue

* special-case null RealmValues in Operator.Convert

* Remove the direct MongoDB.Bson reference from AssemblyToProcess

* Use array pool for string encoding

* Fix peverify reported errors

* Fix warnings

* Fix dynamic bindsetmember

* revert buggy merges

* Use RealmValue/PrimitiveValue to handle objects

* Use the official get/set/find_any on collections

* More meaningful parameter for build.ps1

* More updates

* Fix the unity packager, replace System.Memory with System.Buffers

* Try to fix weaver tests

* Try to fix the compact tests

* More updates for the unity bundler

* Make RealmValue->something operators explicit

* Use latest core

* Use core develop, not master

* Fix compilation errors

+++ wrappers/build.ps1
@@ -20,7 +20,9 @@ param(
 
     [ValidateSet('Windows', 'WindowsStore')]
     [Parameter(Position=0)]
-    [string]$Target = 'Windows'
+    [string]$Target = 'Windows',
+
+    [Switch]$Incremental
 )
 
 Push-Location $PSScriptRoot
@@ -61,10 +63,15 @@ function triplet([string]$target, [string]$platform) {
 }
 
 foreach ($platform in $Platforms) {
-    Remove-Item .\cmake\$Target\$Configuration-$platform -Recurse -Force -ErrorAction Ignore
+    if (-Not $Incremental) {
+        Remove-Item .\cmake\$Target\$Configuration-$platform -Recurse -Force -ErrorAction Ignore
+    }
     New-Item .\cmake\$Target\$Configuration-$platform -ItemType "Directory" | Out-Null
     Push-Location .\cmake\$Target\$Configuration-$platform
-    & $cmake $PSScriptRoot $cmakeArgs -DCMAKE_GENERATOR_PLATFORM="$platform" -DVCPKG_TARGET_TRIPLET="$(triplet -target $Target -platform $platform)"
+    if (-Not $Incremental) {
+        & $cmake $PSScriptRoot $cmakeArgs -DCMAKE_GENERATOR_PLATFORM="$platform" -DVCPKG_TARGET_TRIPLET="$(triplet -target $Target -platform $platform)"
+    }
+    
     & $cmake --build . --target install --config $Configuration
     Pop-Location
 }
